# swagger.yaml
openapi: 3.0.0
info:
  title: Node.js Sign-up and Verification API
  description: Node.js + MySQL - API with email sign-up, verification, authentication and forgot password
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
  /accounts/authenticate:
    post:
      summary: Authenticate account credentials and return a JWT token and a cookie with a refresh token
      operationId: authenticate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: admin@example.com
                password:
                  type: string
                  example: Asdf1234
              required:
                - email
                - password
      responses:
        '200':
          description: Account details, JWT token and refresh token cookie
  /accounts/refresh-token:
    post:
      summary: Use a refresh token to generate a new JWT token and a new refresh token
      operationId: refreshToken
      responses:
        '200':
          description: Account details, new JWT token and new refresh token cookie
  /accounts/revoke-token:
    post:
      summary: Revoke a refresh token
      operationId: revokeToken
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Token revoked
  /accounts/register:
    post:
      summary: Register a new user account and send a verification email
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Mr
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
                email:
                  type: string
                  example: john@example.com
                password:
                  type: string
                  example: Asdf1234
                confirmPassword:
                  type: string
                  example: Asdf1234
                acceptTerms:
                  type: boolean
                  example: true
              required:
                - title
                - firstName
                - lastName
                - email
                - password
                - confirmPassword
                - acceptTerms
      responses:
        '200':
          description: Registration successful
  /accounts/verify-email:
    post:
      summary: Verify a new account with a verification token received by email after registration
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
              required:
                - token
      responses:
        '200':
          description: Verification successful
  /accounts/forgot-password:
    post:
      summary: Submit email address to reset the password on an account
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: admin@example.com
              required:
                - email
      responses:
        '200':
          description: Password reset email sent
  /accounts/validate-reset-token:
    post:
      summary: Validate the reset password token received by email
      operationId: validateResetToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
              required:
                - token
      responses:
        '200':
          description: Token is valid
  /accounts/reset-password:
    post:
      summary: Reset the password for an account
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                password:
                  type: string
                  example: newPassword123
                confirmPassword:
                  type: string
                  example: newPassword123
              required:
                - token
                - password
                - confirmPassword
      responses:
        '200':
          description: Password reset successful
  /accounts:
    get:
      summary: Get a list of all accounts (Admin only)
      operationId: getAllAccounts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of all accounts
    post:
      summary: Create a new account (Admin only)
      operationId: createAccount
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Account created
  /accounts/{id}:
    get:
      summary: Get a single account by id
      operationId: getAccountById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Account details
    put:
      summary: Update an account
      operationId: updateAccount
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Account updated
    delete:
      summary: Delete an account
      operationId: deleteAccount
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Account deleted

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT